ref(shared): relocate leaky cross-module value objects
Move NamespaceInformation from Build\Domain\Extractor to Phel\Shared so
the Build/Run facade interfaces no longer back-reference a foreign module's
Domain. Source CompilerFacadeInterface::lexString's DEFAULT_SOURCE default
from Shared\CompilerConstants instead of the Application\Lexer concrete, and
sanction EvalError as part of RunFacadeInterface's EvalResult value graph.
HttpClient kept as-is (interop-only callers; Shared forbids I/O utilities).

Related to #2261
