+++ src/php/Run/Application/FileRunner.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Phel\Run\Application;
 
-use Phel\Build\Domain\Extractor\NamespaceInformation;
 use Phel\Compiler\Domain\Analyzer\Resolver\LoadClasspath;
 use Phel\Shared\CompilerConstants;
 use Phel\Shared\Facade\BuildFacadeInterface;
 use Phel\Shared\Facade\CommandFacadeInterface;
+use Phel\Shared\NamespaceInformation;
 use Throwable;
 
 use function array_unique;
