+++ src/php/Run/Infrastructure/Command/TestCommand.php
@@ -7,7 +7,6 @@ namespace Phel\Run\Infrastructure\Command;
 use Gacela\Framework\ServiceResolver\ServiceMap;
 use Gacela\Framework\ServiceResolverAwareTrait;
 use InvalidArgumentException;
-use Phel\Build\Domain\Extractor\NamespaceInformation;
 use Phel\Lang\ProfilerHookInterface;
 use Phel\Lang\Registry;
 use Phel\Run\Domain\Test\TestCommandOptions;
@@ -15,6 +14,7 @@ use Phel\Run\Domain\Test\TestNamespacePruner;
 use Phel\Run\RunFacade;
 use Phel\Shared\CompileOptions;
 use Phel\Shared\Exceptions\CompilerException;
+use Phel\Shared\NamespaceInformation;
 use Phel\Shared\PhelProjectDirectory;
 use Phel\Shared\ResourceUsageFormatter;
 use Symfony\Component\Console\Command\Command;
