+++ src/php/Shared/CompilerConstants.php
@@ -7,4 +7,6 @@ namespace Phel\Shared;
 interface CompilerConstants
 {
     public const string PHEL_CORE_NAMESPACE = 'phel.core';
+
+    public const string DEFAULT_SOURCE = 'string';
 }
