+++ src/php/Shared/Facade/RunFacadeInterface.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Phel\Shared\Facade;
 
-use Phel\Build\Domain\Extractor\NamespaceInformation;
 use Phel\Run\Domain\Repl\EvalResult;
 use Phel\Shared\CompileOptions;
 use Phel\Shared\Exceptions\CompilerException;
+use Phel\Shared\NamespaceInformation;
 use Symfony\Component\Console\Output\OutputInterface;
 use Throwable;
 
@@ -25,6 +25,9 @@ interface RunFacadeInterface
     /**
      * Evaluate Phel code returning a structured result with error details.
      * Unlike eval(), this never throws — errors are captured in EvalResult.
+     *
+     * The returned {@see EvalResult} carries the failure details as an
+     * {@see EvalError}; both are sanctioned parts of this contract's value graph.
      */
     public function structuredEval(string $phelCode, CompileOptions $compileOptions): EvalResult;
