+++ tests/php/Unit/Run/Application/BundledNamespacesTest.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace PhelTest\Unit\Run\Application;
 
-use Phel\Build\Domain\Extractor\NamespaceInformation;
 use Phel\Run\Application\BundledNamespaces;
 use Phel\Shared\Facade\BuildFacadeInterface;
 use Phel\Shared\Facade\CommandFacadeInterface;
+use Phel\Shared\NamespaceInformation;
 use PHPUnit\Framework\TestCase;
 
 final class BundledNamespacesTest extends TestCase
