+++ tests/php/Unit/Run/Application/DataReadersLoaderTest.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace PhelTest\Unit\Run\Application;
 
 use Phel\Build\Domain\Compile\CompiledFile;
-use Phel\Build\Domain\Extractor\NamespaceInformation;
 use Phel\Run\Application\DataReadersLoader;
 use Phel\Shared\Facade\BuildFacadeInterface;
+use Phel\Shared\NamespaceInformation;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
