+++ tests/php/Unit/Watch/Application/ReloadOrchestratorTest.php
@@ -7,12 +7,12 @@ namespace PhelTest\Unit\Watch\Application;
 use Gacela\Framework\Health\ModuleHealthCheckInterface;
 use Phel\Build\Domain\Compile\BuildOptions;
 use Phel\Build\Domain\Compile\CompiledFile;
-use Phel\Build\Domain\Extractor\NamespaceInformation;
 use Phel\Run\Domain\Repl\EvalResult;
 use Phel\Shared\CompileOptions;
 use Phel\Shared\Exceptions\CompilerException;
 use Phel\Shared\Facade\BuildFacadeInterface;
 use Phel\Shared\Facade\RunFacadeInterface;
+use Phel\Shared\NamespaceInformation;
 use Phel\Watch\Application\NamespaceResolver;
 use Phel\Watch\Application\ReloadOrchestrator;
 use Phel\Watch\Domain\ProjectReindexerInterface;
